# docs/swagger.yaml
definitions:
  model.AddPlanet:
    properties:
      climate:
        example: arid
        type: string
      name:
        example: Matheus
        type: string
      terrain:
        example: desert
        type: string
    type: object
  model.Film:
    properties:
      director:
        type: string
      episode_id:
        type: integer
      id:
        type: string
      id_reference:
        type: integer
      release_date:
        type: string
      title:
        type: string
    type: object
  model.Planet:
    properties:
      climate:
        type: string
      films:
        items:
          $ref: '#/definitions/model.Film'
        type: array
      id:
        type: string
      name:
        type: string
      terrain:
        type: string
    type: object
info:
  contact:
    email: support@starwarsapi.io
    name: API Support
    url: http://localhost:9000
paths:
  /films:
    get:
      consumes:
      - application/json
      description: Get all films recorded
      parameters:
      - description: Search film by title
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Film'
      summary: Get all films recorded
      tags:
      - film
  /planets:
    get:
      consumes:
      - application/json
      description: Get all planets recorded
      parameters:
      - description: Search film by exactly name
        in: query
        name: name
        type: string
      - description: Search film by exactly id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Planet'
        "404":
          description: Not Found
      summary: Get all planets recorded
      tags:
      - planet
    post:
      consumes:
      - application/json
      description: Insert a new planet
      parameters:
      - description: Add Movie. Send without the key 'films'.
        in: body
        name: planet
        required: true
        schema:
          $ref: '#/definitions/model.AddPlanet'
      produces:
      - application/json
      responses:
        "201":
          description: a
          schema:
            $ref: '#/definitions/model.Planet'
      summary: Insert a new planet
      tags:
      - planet
  /planets/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a planet by ObjectID
      parameters:
      - description: Planet Object ID to remove
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: Registro
        "400":
          description: Bad Request
          schema:
            type: ObjectID
        "401":
          description: Unauthorized
          schema:
            type: ObjectID
      summary: Remove a planet by ObjectID
      tags:
      - planet
swagger: "2.0"